pyproject.toml
[project]
name = "spotify-playlist-shuffle"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "spotipy>=2.23.0",
    "python-dotenv>=1.0.0"
]
